src/edge_ai/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
llama-cpp-python>=0.2.0
asyncio>=3.4.3
